Add support for symfony 5 event dispatcher
* Ensure BC with old dispatcher (<=4.4)
* Support optional dispatcher API (Symfony >=5)
* Update readme
* Update Release file
