+++ RELEASE.md
@@ -1,5 +1,10 @@
 # Release notes
 
+# 3.0.0
+
+* Update composer dependencies
+* Add support for symfony event dispatcher api >-5.0
+
 # 2.0.0
 
 * Upgrade/php version bump (**BC break**).
